4.22*: bloglist expansion, step10

+++ blog-list/controllers/blogs.js
@@ -1,15 +1,9 @@
 const blogsRouter = require("express").Router();
-// const jwt = require('jsonwebtoken')
+const jwt = require("jsonwebtoken");
+
 const Blog = require("../models/blog");
 const User = require("../models/user");
-
-const getTokenFrom = (request) => {
-  const authorization = request.get("authorization");
-  if (authorization && authorization.toLowerCase().startsWith("bearer")) {
-    return authorization.substring(7);
-  }
-  return null;
-};
+const logger = require("../utils/logger");
 
 blogsRouter.get("/", async (request, response) => {
   const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -18,9 +12,15 @@ blogsRouter.get("/", async (request, response) => {
 });
 
 blogsRouter.post("/", async (request, response) => {
-  let body = request.body;
+  const body = request.body;
 
-  const user = await User.findById(body.userId);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const user = await User.findById(decodedToken.id);
 
   if (!user) {
     return response.status(404).json({
@@ -45,6 +45,18 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response, next) => {
+  const blogToDelete = await Blog.findById(request.params.id);
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (
+    blogToDelete &&
+    decodedToken.id &&
+    blogToDelete.user.toString() !== decodedToken.id.toString()
+  ) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
   await Blog.findByIdAndRemove(request.params.id);
   response.status(204).end();
 });
